requirements.txt
Flask==2.3.3
openai==0.28.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7
flask-cors==4.0.0
requests==2.31.0
urllib3==2.0.7
pinecone
sentence-transformers==2.2.2
numpy>=1.21.0
torch>=1.9.0
transformers>=4.21.0 